backend/requirements.txt
flask==3.1.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
opencv-python==4.9.0.80
numpy==1.26.4
werkzeug==3.1.3
gunicorn==22.0.0
python-dotenv==1.0.1
Pillow==10.2.0
PyJWT==2.12.1
